CNAM-361: refactor consecutive falls
1: refactor the group of consecutive falls
2: optimise the package
