+++ src/main/scala/fr/polytechnique/cmap/cnam/study/fall/fractures/FracturesTransformer.scala
@@ -1,11 +1,11 @@
-package fr.polytechnique.cmap.cnam.study.fall
+package fr.polytechnique.cmap.cnam.study.fall.fractures
 
 import org.apache.spark.sql.Dataset
 import fr.polytechnique.cmap.cnam.etl.events._
 import fr.polytechnique.cmap.cnam.etl.transformers.outcomes.OutcomesTransformer
-import fr.polytechnique.cmap.cnam.study.fall.FracturesTransformerImplicits._
 import fr.polytechnique.cmap.cnam.study.fall.codes.FractureCodes
-import fr.polytechnique.cmap.cnam.study.fall.config.{FallConfig, FracturesTransformerConfig}
+import fr.polytechnique.cmap.cnam.study.fall.config.FallConfig
+import fr.polytechnique.cmap.cnam.study.fall.fractures.FracturesTransformerImplicits._
 import fr.polytechnique.cmap.cnam.util.functions.unionDatasets
 
 /*
