+++ src/main/scala/fr/polytechnique/cmap/cnam/study/fall/fractures/LiberalFractures.scala
@@ -1,4 +1,4 @@
-package fr.polytechnique.cmap.cnam.study.fall
+package fr.polytechnique.cmap.cnam.study.fall.fractures
 
 import fr.polytechnique.cmap.cnam.etl.events.{Event, MedicalAct, Outcome}
 import fr.polytechnique.cmap.cnam.etl.transformers.outcomes.OutcomesTransformer
