+++ src/test/scala/fr/polytechnique/cmap/cnam/study/fall/fractures/HospitalizedFracturesSuite.scala
@@ -2,7 +2,6 @@ package fr.polytechnique.cmap.cnam.study.fall.fractures
 
 import fr.polytechnique.cmap.cnam.SharedContext
 import fr.polytechnique.cmap.cnam.etl.events._
-import fr.polytechnique.cmap.cnam.study.fall.{AllSites, HospitalStay, HospitalizedFractures}
 import fr.polytechnique.cmap.cnam.util.functions._
